Move dergon list command logic out of dergonHandler and into the command class
